refactor: giving a better specs
